Trait oasis_runtime_sdk::context::Context

pub trait Context {
    type Runtime: Runtime;
    type Store: Store;

    fn get_logger(&self, module: &'static str) -> Logger;

    fn mode(&self) -> Mode;

    fn host_info(&self) -> &HostInfo;

    fn runtime_header(&self) -> &Header;

    fn runtime_round_results(&self) -> &RoundResults;

    fn runtime_state(&mut self) -> &mut Self::Store;

    fn consensus_state(&self) -> &ConsensusState;

    fn epoch(&self) -> EpochTime;

    fn emit_event<E: Event>(&mut self, event: E);

    fn io_ctx(&self) -> IoContext;

    fn commit(self) -> (Tags, Vec<(Message, MessageEventHookInvocation)>);

    fn value<V: Any>(&mut self, key: &'static str) -> ContextValue<'_, V>;

    fn with_simulation<F, Rs>(&mut self, f: F) -> Rs
    where
        F: FnOnce(RuntimeBatchContext<'_, Self::Runtime, OverlayStore<&mut Self::Store>>) -> Rs;

    fn is_check_only(&self) -> bool { ... }

    fn is_simulation(&self) -> bool { ... }

    fn runtime_id(&self) -> &Namespace { ... }

}

Runtime SDK context.

Associated Types

type Runtime: Runtime

Runtime that the context is being invoked in.

type Store: Store

Runtime state output type.

Loading content...
Required methods

fn get_logger(&self, module: &'static str) -> Logger

Returns a logger.

fn mode(&self) -> Mode

Context mode.

fn host_info(&self) -> &HostInfo

Information about the host environment.

fn runtime_header(&self) -> &Header

Last runtime block header.

fn runtime_round_results(&self) -> &RoundResults

Results of executing the last successful runtime round.

fn runtime_state(&mut self) -> &mut Self::Store

Runtime state store.

fn consensus_state(&self) -> &ConsensusState

Consensus state.

fn epoch(&self) -> EpochTime

Current epoch.

fn emit_event<E: Event>(&mut self, event: E)

Emits an event.

fn io_ctx(&self) -> IoContext

Returns a child io_ctx.

fn commit(self) -> (Tags, Vec<(Message, MessageEventHookInvocation)>)

Commit any changes made to storage, return any emitted tags and runtime messages. It consumes the transaction context.

fn value<V: Any>(&mut self, key: &'static str) -> ContextValue<'_, V>

Fetches a value entry associated with the context.

fn with_simulation<F, Rs>(&mut self, f: F) -> Rs where
F: FnOnce(RuntimeBatchContext<'_, Self::Runtime, OverlayStore<&mut Self::Store>>) -> Rs,

Executes a function in a simulation context.

The simulation context collects its own messages and starts with an empty set of context values.

Loading content...
Provided methods

fn is_check_only(&self) -> bool

Whether the transaction is just being checked for validity.

fn is_simulation(&self) -> bool

Whether the transaction is just being simulated.

fn runtime_id(&self) -> &Namespace

Runtime ID.

Loading content...