oasis_runtime_sdk::crypto::signature::context

Function get_chain_context_for

pub fn get_chain_context_for(base: &[u8]) -> Vec<u8> 

Return the globally configured chain domain separation context.

The returned domain separation context is computed as:

<base> || " for chain " || <chain-context>

Panics

This function will panic in case the global chain domain separation context was not previously set using set_chain_context.