oasis_runtime_sdk::crypto::signature

Enum MemorySigner

pub enum MemorySigner {
    Ed25519(MemorySigner),
    Secp256k1(MemorySigner),
    Secp256r1(MemorySigner),
    Secp384r1(MemorySigner),
    Sr25519(MemorySigner),
}

A memory-backed signer.

Variants

Ed25519(MemorySigner)

Secp256k1(MemorySigner)

Secp256r1(MemorySigner)

Secp384r1(MemorySigner)

Sr25519(MemorySigner)

Implementations

impl MemorySigner

pub fn new_from_seed( sig_type: SignatureType, seed: &[u8], ) -> Result<Self, Error>

Create a new memory signer from a seed.

pub fn new_test(sig_type: SignatureType, name: &str) -> Self

Create a new signer for testing purposes.

pub fn from_bytes(sig_type: SignatureType, bytes: &[u8]) -> Result<Self, Error>

Reconstruct the signer from its byte representation.

pub fn to_bytes(&self) -> Vec<u8>

Return a byte representation of the signer.

pub fn public_key(&self) -> PublicKey

Public key corresponding to the signer.

pub fn sign(&self, context: &[u8], message: &[u8]) -> Result<Signature, Error>

Generate a signature with the private key over the context and message.

pub fn sign_raw(&self, message: &[u8]) -> Result<Signature, Error>

Generate a signature with the private key over the message.

pub fn sign_by_type( &self, signature_type: SignatureType, context_or_hash: &[u8], message: &[u8], ) -> Result<Signature, Error>

Generate a signature for the specified message and optional context.