Module oasis_runtime_sdk::module

Runtime modules.

Enums

AuthDecision

An authentication decision for cases where multiple handlers are available.

CallResult

A variant of types::transaction::CallResult but used for dispatch purposes so the dispatch process can use a different representation.

DispatchResult

Result of invoking the method handler.

Traits

BlockHandler

Block handler.

FeeProxyHandler

Fee proxy handler.

InvariantHandler

Invariant handler.

MethodHandler

Method handler.

MigrationHandler

Migration handler.

Module

A runtime module.

ModuleInfoHandler

Info handler.

Parameters

Parameters for a runtime module.

TransactionHandler

Transaction handler.

Functions

dispatch_call

A convenience function for dispatching method calls.

dispatch_query

A convenience function for dispatching queries.