Enum oasis_runtime_sdk::module::CallResult

pub enum CallResult {
    Ok(Value),
    Failed {
        module: String,
        code: u32,
        message: String,
    },
    Aborted(Error),
}

A variant of types::transaction::CallResult but used for dispatch purposes so the dispatch process can use a different representation.

Specifically, this type is not serializable.

Variants

Ok(Value)

Call has completed successfully.

Failed

Fields

module: String

code: u32

message: String

Call has completed with failure.

Aborted(Error)

A fatal error has occurred and the batch must be aborted.

Implementations

impl CallResult

pub fn is_success(&self) -> bool

Check whether the call result indicates a successful operation or not.

pub fn unwrap(self) -> Value