Trait oasis_runtime_sdk::module::AuthHandler

pub trait AuthHandler {
    fn approve_unverified_tx<C: Context>(
        _ctx: &mut C, 
        _utx: &UnverifiedTransaction
    ) -> Result<(), Error> { ... }

    fn authenticate_tx<C: Context>(
        _ctx: &mut C, 
        _tx: &Transaction
    ) -> Result<(), Error> { ... }

    fn before_handle_call<C: TxContext>(
        _ctx: &mut C, 
        _call: &Call
    ) -> Result<(), Error> { ... }
}

Authentication handler.

Provided methods

fn approve_unverified_tx<C: Context>(
_ctx: &mut C,
_utx: &UnverifiedTransaction
) -> Result<(), Error>

Judge if an unverified transaction is good enough to undergo verification. This takes place before even verifying signatures.

fn authenticate_tx<C: Context>(
_ctx: &mut C,
_tx: &Transaction
) -> Result<(), Error>

Authenticate a transaction.

Note that any signatures have already been verified.

fn before_handle_call<C: TxContext>(
_ctx: &mut C,
_call: &Call
) -> Result<(), Error>

Perform any action after authentication, within the transaction context.

Loading content...