Trait oasis_runtime_sdk::module::BlockHandler

pub trait BlockHandler {
    fn begin_block<C: Context>(_ctx: &mut C) { ... }

    fn end_block<C: Context>(_ctx: &mut C) { ... }

    fn get_block_weight_limits<C: Context>(
        _ctx: &mut C
    ) -> BTreeMap<TransactionWeight, u64> { ... }
}

Block handler.

Provided methods

fn begin_block<C: Context>(_ctx: &mut C)

Perform any common actions at the start of the block (before any transactions have been executed).

fn end_block<C: Context>(_ctx: &mut C)

Perform any common actions at the end of the block (after all transactions have been executed).

fn get_block_weight_limits<C: Context>(
_ctx: &mut C
) -> BTreeMap<TransactionWeight, u64>

Returns module per-batch weight limits.

Loading content...