Trait oasis_runtime_sdk::module::MigrationHandler

pub trait MigrationHandler {
    type Genesis;

    // Provided method
    fn init_or_migrate<C: Context>(
        _ctx: &mut C,
        _meta: &mut Metadata,
        _genesis: Self::Genesis
    ) -> bool { ... }
}

Migration handler.

Required Associated Types

type Genesis

Genesis state type.

If this state is expensive to compute and not often updated, prefer to make the genesis type something like once_cell::unsync::Lazy<T>.

Provided Methods

fn init_or_migrate<C: Context>( _ctx: &mut C, _meta: &mut Metadata, _genesis: Self::Genesis ) -> bool

Initialize state from genesis or perform a migration.

Should return true in case metadata has been changed.