Module oasis_runtime_sdk::modules::accounts::fee

Fee manager.

Structs

FeeManager

The per-block fee manager that records what fees have been charged by the current transaction, how much should be refunded and what were all of the fee payments in the current block.

FeeUpdates

Fee updates to apply to state after commit_tx.

TransactionFee

Information about fees charged for the current transaction.