Trait oasis_runtime_sdk::modules::accounts::API

pub trait API {

    fn transfer<C: Context>(
        ctx: &mut C,
        from: Address,
        to: Address,
        amount: &BaseUnits
    ) -> Result<(), Error>;
    fn mint<C: Context>(
        ctx: &mut C,
        to: Address,
        amount: &BaseUnits
    ) -> Result<(), Error>;
    fn burn<C: Context>(
        ctx: &mut C,
        from: Address,
        amount: &BaseUnits
    ) -> Result<(), Error>;
    fn set_nonce<S: Store>(state: S, address: Address, nonce: u64);
    fn get_nonce<S: Store>(state: S, address: Address) -> Result<u64, Error>;
    fn set_balance<S: Store>(state: S, address: Address, amount: &BaseUnits);
    fn get_balance<S: Store>(
        state: S,
        address: Address,
        denomination: Denomination
    ) -> Result<u128, Error>;
    fn get_balances<S: Store>(
        state: S,
        address: Address
    ) -> Result<AccountBalances, Error>;
    fn get_addresses<S: Store>(
        state: S,
        denomination: Denomination
    ) -> Result<Vec<Address>, Error>;
    fn get_total_supplies<S: Store>(
        state: S
    ) -> Result<BTreeMap<Denomination, u128>, Error>;
    fn set_total_supply<S: Store>(state: S, amount: &BaseUnits);
    fn get_denomination_info<S: Store>(
        state: S,
        denomination: &Denomination
    ) -> Result<DenominationInfo, Error>;
    fn charge_tx_fee<C: Context>(
        ctx: &mut C,
        from: Address,
        amount: &BaseUnits
    ) -> Result<(), Error>;
    fn return_tx_fee<C: Context>(
        ctx: &mut C,
        to: Address,
        amount: &BaseUnits
    ) -> Result<(), Error>;
    fn check_signer_nonces<C: Context>(
        ctx: &mut C,
        tx_auth_info: &AuthInfo
    ) -> Result<Address, Error>;
    fn update_signer_nonces<C: Context>(
        ctx: &mut C,
        tx_auth_info: &AuthInfo
    ) -> Result<(), Error>;

    fn ensure_balance<S: Store>(
        state: S,
        address: Address,
        amount: &BaseUnits
    ) -> Result<(), Error> { ... }

}

Interface that can be called from other modules.

Required Methods

fn transfer<C: Context>(
ctx: &mut C,
from: Address,
to: Address,
amount: &BaseUnits
) -> Result<(), Error>

Transfer an amount from one account to the other.

fn mint<C: Context>(
ctx: &mut C,
to: Address,
amount: &BaseUnits
) -> Result<(), Error>

Mint new tokens, increasing the total supply.

fn burn<C: Context>(
ctx: &mut C,
from: Address,
amount: &BaseUnits
) -> Result<(), Error>

Burn existing tokens, decreasing the total supply.

fn set_nonce<S: Store>(state: S, address: Address, nonce: u64)

Sets an account’s nonce.

fn get_nonce<S: Store>(state: S, address: Address) -> Result<u64, Error>

Fetch an account’s current nonce.

fn set_balance<S: Store>(state: S, address: Address, amount: &BaseUnits)

Sets an account’s balance of the given denomination.

Warning

This method is dangerous as it can result in invariant violations.

fn get_balance<S: Store>(
state: S,
address: Address,
denomination: Denomination
) -> Result<u128, Error>

Fetch an account’s balance of the given denomination.

fn get_balances<S: Store>(
state: S,
address: Address
) -> Result<AccountBalances, Error>

Fetch an account’s current balances.

fn get_addresses<S: Store>(
state: S,
denomination: Denomination
) -> Result<Vec<Address>, Error>

Fetch addresses.

fn get_total_supplies<S: Store>(
state: S
) -> Result<BTreeMap<Denomination, u128>, Error>

Fetch total supplies.

fn set_total_supply<S: Store>(state: S, amount: &BaseUnits)

Sets the total supply for the given denomination.

Warning

This method is dangerous as it can result in invariant violations.

fn get_denomination_info<S: Store>(
state: S,
denomination: &Denomination
) -> Result<DenominationInfo, Error>

Fetch information about a denomination.

fn charge_tx_fee<C: Context>(
ctx: &mut C,
from: Address,
amount: &BaseUnits
) -> Result<(), Error>

Moves the amount into the per-transaction fee accumulator.

fn return_tx_fee<C: Context>(
ctx: &mut C,
to: Address,
amount: &BaseUnits
) -> Result<(), Error>

Returns the amount from the per-transaction fee accumulator to the account.

fn check_signer_nonces<C: Context>(
ctx: &mut C,
tx_auth_info: &AuthInfo
) -> Result<Address, Error>

Check transaction signer account nonces. Return payer address.

fn update_signer_nonces<C: Context>(
ctx: &mut C,
tx_auth_info: &AuthInfo
) -> Result<(), Error>

Update transaction signer account nonces.

Provided Methods

fn ensure_balance<S: Store>(
state: S,
address: Address,
amount: &BaseUnits
) -> Result<(), Error>

Ensures that the given account has at least the specified balance.