Trait oasis_runtime_sdk::modules::core::API

pub trait API {
    fn use_batch_gas<C: Context>(ctx: &mut C, gas: u64) -> Result<(), Error>;

    fn use_tx_gas<C: TxContext>(ctx: &mut C, gas: u64) -> Result<(), Error>;

    fn add_priority<C: Context>(ctx: &mut C, priority: u64) -> Result<(), Error>;

    fn add_weight<C: TxContext>(
        ctx: &mut C, 
        weight: TransactionWeight, 
        val: u64
    ) -> Result<(), Error>;

    fn take_weights<C: Context>(ctx: &mut C) -> BTreeMap<TransactionWeight, u64>;

    fn take_priority<C: Context>(ctx: &mut C) -> u64;
}

Required methods

fn use_batch_gas<C: Context>(ctx: &mut C, gas: u64) -> Result<(), Error>

Attempt to use gas. If the gas specified would cause either total used to exceed its limit, fails with Error::OutOfGas or Error::BatchOutOfGas, and neither gas usage is increased.

fn use_tx_gas<C: TxContext>(ctx: &mut C, gas: u64) -> Result<(), Error>

Attempt to use gas. If the gas specified would cause either total used to exceed its limit, fails with Error::OutOfGas or Error::BatchOutOfGas, and neither gas usage is increased.

fn add_priority<C: Context>(ctx: &mut C, priority: u64) -> Result<(), Error>

Increase transaction priority for the provided amount.

fn add_weight<C: TxContext>(
ctx: &mut C,
weight: TransactionWeight,
val: u64
) -> Result<(), Error>

Increase the specific transaction weight for the provided amount.

fn take_weights<C: Context>(ctx: &mut C) -> BTreeMap<TransactionWeight, u64>

Takes the stored transaction weight.

fn take_priority<C: Context>(ctx: &mut C) -> u64

Takes and returns the stored transaction priority.

Loading content...