oasis_runtime_sdk::modules::rofl::app::client

Struct Client

pub struct Client<A: App> { /* private fields */ }

A runtime client meant for use within runtimes.

Implementations

impl<A> Client<A>
where A: App,

pub async fn latest_round(&self) -> Result<u64>

Retrieve the latest known runtime round.

pub async fn account_nonce(&self, round: u64, address: Address) -> Result<u64>

Retrieve the nonce for the given account.

pub async fn gas_price(&self, round: u64, denom: &Denomination) -> Result<u128>

Retrieve the gas price in the given denomination.

pub async fn query<Rq, Rs>( &self, round: u64, method: &str, args: Rq, ) -> Result<Rs>
where Rq: Encode, Rs: Decode + Send + 'static,

Securely query the on-chain runtime component.

pub async fn estimate_gas(&self, req: EstimateGasQuery) -> Result<u64>

Securely perform gas estimation.

pub async fn app_cfg(&self) -> Result<AppConfig>

Retrieves application configuration.

pub async fn multi_sign_and_submit_tx( &self, signers: &[Arc<dyn Signer>], tx: Transaction, ) -> Result<CallResult>

Sign a given transaction, submit it and wait for block inclusion.

This method supports multiple transaction signers.

pub async fn multi_sign_and_submit_tx_opts( &self, signers: &[Arc<dyn Signer>], tx: Transaction, opts: SubmitTxOpts, ) -> Result<CallResult>

Sign a given transaction, submit it and wait for block inclusion.

This method supports multiple transaction signers.

pub async fn sign_and_submit_tx( &self, signer: Arc<dyn Signer>, tx: Transaction, ) -> Result<CallResult>

Sign a given transaction, submit it and wait for block inclusion.

pub async fn with_store_for_round<F, R>(&self, round: u64, f: F) -> Result<R>
where F: FnOnce() -> Result<R> + Send + 'static, R: Send + 'static,

Run a closure inside a CurrentState context with store for the given round.

pub async fn store_for_round(&self, round: u64) -> Result<HostStore>

Return a store corresponding to the given round.

pub async fn derive_key( &self, signer: Arc<dyn Signer>, request: DeriveKeyRequest, ) -> Result<DeriveKeyResponse>

Derive an application-specific key.