oasis_runtime_sdk::modules::rofl

Module policy

Structs

AppAuthPolicy

Per-application ROFL policy.

BasicEndorsementPolicyEvaluator

An endorsement policy evaluator which implements support for basic endorsement atoms.

Enums

AllowedEndorsement

An allowed endorsement policy.

EndorsementPolicyOperator

Endorsement policy operator.

FeePolicy

Gas fee payment policy.

Constants

MAX_ENDORSEMENT_POLICY_DEPTH

Maximum depth when evaluating an endorsement policy.

Traits

EndorsementPolicyEvaluator

An evaluator of an endorsement policy which decides whether the node that endorsed the given enclave is acceptable accoording to a policy.