Trait oasis_runtime_sdk::modules::rofl::Config

pub trait Config: 'static {
    const GAS_COST_CALL_CREATE: u64 = 100_000u64;
    const GAS_COST_CALL_UPDATE: u64 = 100_000u64;
    const GAS_COST_CALL_REMOVE: u64 = 10_000u64;
    const GAS_COST_CALL_REGISTER: u64 = 100_000u64;
    const GAS_COST_CALL_IS_AUTHORIZED_ORIGIN: u64 = 1_000u64;
    const STAKE_APP_CREATE: BaseUnits = _;
}

Module configuration.

Provided Associated Constants

const GAS_COST_CALL_CREATE: u64 = 100_000u64

Gas cost of rofl.Create call.

const GAS_COST_CALL_UPDATE: u64 = 100_000u64

Gas cost of rofl.Update call.

const GAS_COST_CALL_REMOVE: u64 = 10_000u64

Gas cost of rofl.Remove call.

const GAS_COST_CALL_REGISTER: u64 = 100_000u64

Gas cost of rofl.Register call.

const GAS_COST_CALL_IS_AUTHORIZED_ORIGIN: u64 = 1_000u64

Gas cost of rofl.IsAuthorizedOrigin call.

const STAKE_APP_CREATE: BaseUnits = _

Amount of stake required for maintaining an application.

The stake is held in escrow and is returned to the administrator when the application is removed.

Object Safety

This trait is not object safe.