oasis_runtime_sdk::state

Struct CurrentState

pub struct CurrentState;

State attached to the current thread.

Implementations

impl CurrentState

pub fn enter<S, F, R>(root: S, f: F) -> R
where S: Store, F: FnOnce() -> R,

Attach a new state to the current thread and enter the state’s context.

The passed store is used as the root store.

Panics

This method will panic if called from within a CurrentState::with block.

pub fn enter_opts<S, F, R>(opts: Options, root: S, f: F) -> R
where S: Store, F: FnOnce() -> R,

Attach a new state to the current thread and enter the state’s context.

The passed store is used as the root store.

Panics

This method will panic if called from within a CurrentState::with block or if the mode has not been explicitly set in opts.

pub fn with<F, R>(f: F) -> R
where F: FnOnce(&mut State) -> R,

Run a closure with the currently active state.

Panics

This method will panic if called outside CurrentState::enter or if any transaction methods are called from the closure.

pub fn with_store<F, R>(f: F) -> R
where F: FnOnce(&mut dyn Store) -> R,

Run a closure with the store of the currently active state.

Panics

This method will panic if called outside CurrentState::enter or if any transaction methods are called from the closure.

pub fn with_env<F, R>(f: F) -> R
where F: FnOnce(&Environment) -> R,

Run a closure with the environment of the currently active state.

Panics

This method will panic if called outside CurrentState::enter or if any transaction methods are called from the closure.

pub fn with_env_origin<F, R>(f: F) -> R
where F: FnOnce(&Environment) -> R,

Run a closure with the origin environment of the currently active state.

Panics

This method will panic if called outside CurrentState::enter or if any transaction methods are called from the closure.

pub fn start_transaction()

Start a new transaction by opening a new child state.

Panics

This method will panic if called outside CurrentState::enter or if called within a CurrentState::with block.

pub fn commit_transaction()

Commit a previously started transaction.

Panics

This method will panic if called outside CurrentState::enter, if there is no currently open transaction (started via CurrentState::start_transaction) or if called within a CurrentState::with block.

pub fn rollback_transaction()

Rollback a previously started transaction.

Panics

This method will panic if called outside CurrentState::enter, if there is no currently open transaction (started via CurrentState::start_transaction) or if called within a CurrentState::with block.

pub fn with_transaction<F, R, Rs>(f: F) -> R
where F: FnOnce() -> Rs, Rs: Into<TransactionResult<R>>,

Run a closure within a state transaction.

If the closure returns TransactionResult::Commit(R) then the child state is committed, otherwise the child state is rolled back.

pub fn with_transaction_opts<F, R, Rs>(opts: Options, f: F) -> R
where F: FnOnce() -> Rs, Rs: Into<TransactionResult<R>>,

Run a closure within a state transaction, allowing the caller to customize state.

If the closure returns TransactionResult::Commit(R) then the child state is committed, otherwise the child state is rolled back.