Struct oasis_runtime_sdk::storage::current::CurrentStore

pub struct CurrentStore { /* private fields */ }

A store attached to the current thread.

Implementations

impl CurrentStore

pub fn enter<S, F, R>(root: S, f: F) -> Rwhere S: Store, F: FnOnce() -> R,

Attach a new store to the current thread and enter the store’s context.

The passed store is used as the root store.

pub fn start_transaction() -> usize

Start a new transaction by overlaying a store over the current store.

Panics

This method will panic if called outside CurrentStore::enter or if called within a CurrentStore::with block.

pub fn commit_transaction()

Commit a previously started transaction.

Panics

This method will panic if called outside CurrentStore::enter, if there is no currently open transaction (started via CurrentStore::start_transaction) or if called within a CurrentStore::with block.

pub fn rollback_transaction()

Rollback a previously started transaction.

Panics

This method will panic if called outside CurrentStore::enter, if there is no currently open transaction (started via CurrentStore::start_transaction) or if called within a CurrentStore::with block.

pub fn has_pending_updates() -> bool

Whether there are any store updates pending to be committed in the current transaction.

If there is no current transaction, the method returns true.

Panics

This method will panic if called outside CurrentStore::enter or if called within a CurrentStore::with block.

pub fn with<F, R>(f: F) -> Rwhere F: FnOnce(&mut dyn Store) -> R,

Run a closure with the currently active store.

Panics

This method will panic if called outside CurrentStore::enter or if any transaction methods are called from the closure.

pub fn with_transaction<F, R>(f: F) -> Rwhere F: FnOnce() -> TransactionResult<R>,

Run a closure within a storage transaction.

If the closure returns TransactionResult::Commit(R) then the transaction is committed, otherwise the transaction is rolled back.