Trait oasis_runtime_sdk::storage::NestedStore

pub trait NestedStore: Store {
    type Inner;

    fn commit(self) -> Self::Inner;
    fn has_pending_updates(&self) -> bool;
}

A key-value store that supports the commit operation.

Required Associated Types

type Inner

Type of the inner store.

Required Methods

fn commit(self) -> Self::Inner

Commit any changes to the underlying store.

If this method is not called the changes may be discarded by the store.

fn has_pending_updates(&self) -> bool

Whether there are any store updates pending to be committed.