Trait oasis_runtime_sdk::storage::Store

pub trait Store {
    // Required methods
    fn get(&self, key: &[u8]) -> Option<Vec<u8>>;
    fn insert(&mut self, key: &[u8], value: &[u8]);
    fn remove(&mut self, key: &[u8]);
    fn iter(&self) -> Box<dyn Iterator + '_>;
    fn prefetch_prefixes(&mut self, prefixes: Vec<Prefix>, limit: u16);
}

A key-value store.

Required Methods

fn get(&self, key: &[u8]) -> Option<Vec<u8>>

Fetch entry with given key.

fn insert(&mut self, key: &[u8], value: &[u8])

Update entry with given key to the given value.

fn remove(&mut self, key: &[u8])

Remove entry with given key.

fn iter(&self) -> Box<dyn Iterator + '_>

Returns an iterator over the tree.

fn prefetch_prefixes(&mut self, prefixes: Vec<Prefix>, limit: u16)

Populate the in-memory tree with nodes for keys starting with given prefixes.