Struct oasis_runtime_sdk::types::address::Address

pub struct Address(_);

An account address.

Implementations

impl Address

pub fn new(ctx: &'static [u8], version: u8, data: &[u8]) -> Self

Creates a new address from a context, version and data.

pub fn from_bytes(data: &[u8]) -> Result<Self, Error>

Tries to create a new address from raw bytes.

pub fn from_module(module: &'static str, kind: &'static str) -> Self

Creates a new address for a specific module.

pub fn from_pk(pk: &PublicKey) -> Self

Creates a new address from a public key.

pub fn from_bech32(data: &str) -> Result<Self, Error>

Tries to create a new address from Bech32-encoded string.

pub fn to_bech32(&self) -> String

Converts an address to Bech32 representation.