Crate oasis_runtime_sdk

Oasis runtime SDK.

Re-exports

pub use crate::context::Context;

pub use crate::module::Module;

pub use crate::runtime::Runtime;

pub use crate::state::CurrentState;

pub use oasis_core_runtime as core;

pub use cbor;

Modules

callformat

Handling of different call formats.

config

Configuration types.

context

Execution context.

crypto

Cryptography.

dispatcher

Transaction dispatcher.

enclave_rpc

Exposed EnclaveRPC methods.

error

Error types for runtimes.

event

Event types for runtimes.

history

Historic state access.

keymanager

Keymanager interface.

module

Runtime modules.

modules

Runtime modules included with the SDK.

runtime

Runtime.

schedule_control

Types related to schedule control.

sender

Transaction sender metadata.

state

storage

Storage.

subcall

Subcall dispatch.

testing

Module which contains utilities useful for testing and development.

types

Types defined by the SDK.

Macros

version_from_cargo

Constructs an oasis_sdk::core::common::version::Version from the Cargo.toml version.

Structs

Version

A protocol or runtime version.

Attribute Macros

handler

A helper attribute for #[sdk_derive(...)]. It doesn’t do anyting on its own; it only marks functions that represent a paratime method handler. The permitted forms are:

migration

A helper attribute for #[sdk_derive(...)]. It doesn’t do anything on its own; it only marks functions that represent a module state migration.

sdk_derive

Derives traits from a non-trait impl block (rather than from a struct).

Derive Macros

Error

Derives the Error trait on an enum.

Event

Derives the Event trait on an enum.