Crate oasis_runtime_sdk

Oasis runtime SDK.

Re-exports

pub use crate::context::BatchContext;

pub use crate::context::Context;

pub use crate::context::TxContext;

pub use crate::module::Module;

pub use crate::runtime::Runtime;

pub use oasis_core_runtime as core;

pub use cbor;

Modules

callformat

Handling of different call formats.

config

Configuration types.

context

Execution context.

crypto

Cryptography.

dispatcher

Transaction dispatcher.

error

Error types for runtimes.

event

Event types for runtimes.

keymanager

Keymanager interface.

module

Runtime modules.

modules

Runtime modules included with the SDK.

runtime

Runtime.

schedule_control

Types related to schedule control.

storage

Storage.

testing

Module which contains utilities useful for testing and development.

types

Types defined by the SDK.

Macros

version_from_cargo

Constructs an oasis_sdk::core::common::version::Version from the Cargo.toml version.

Structs

Version

A protocol or runtime version.

Attribute Macros

handler

A helper attribute for #[sdk_derive(...)]. It doesn’t do anyting on its own; it only mark functions that represent a paratime method handler. The permitted forms are:

sdk_derive

Derives traits from a non-trait impl block (rather than from a struct).

Derive Macros

Error

Derives the Error trait on an enum.

Event

Derives the Event trait on an enum.