Trait oasis_runtime_sdk::context::Context

pub trait Context {
    type S: Store;

    fn get_logger(&self, module: &'static str) -> Logger;

    fn mode(&self) -> Mode;

    fn runtime_header(&self) -> &Header;

    fn runtime_round_results(&self) -> &RoundResults;

    fn runtime_state(&mut self) -> &mut Self::S;

    fn consensus_state(&self) -> &ConsensusState;

    fn epoch(&self) -> EpochTime;

    fn tx_auth_info(&self) -> Option<&AuthInfo>;

    fn emit_event<E: Event>(&mut self, event: E);

    fn emit_message(
        &mut self, 
        msg: Message, 
        hook: MessageEventHookInvocation
    ) -> Result<(), Error>;

    fn io_ctx(&self) -> IoContext;

    fn commit(self) -> (Tags, Vec<(Message, MessageEventHookInvocation)>);

    fn value<V: Any>(&mut self, key: &'static str) -> ContextValue<'_, V>;

    fn is_check_only(&self) -> bool { ... }

    fn is_simulation(&self) -> bool { ... }

    fn tx_caller_address(&self) -> Option<Address> { ... }

    fn emit_messages(
        &mut self, 
        msgs: Vec<(Message, MessageEventHookInvocation)>
    ) -> Result<(), Error> { ... }

}

Runtime SDK context.

Associated Types

type S: Store

Runtime state output type.

Loading content...
Required methods

fn get_logger(&self, module: &'static str) -> Logger

Returns a logger.

fn mode(&self) -> Mode

Context mode.

fn runtime_header(&self) -> &Header

Last runtime block header.

fn runtime_round_results(&self) -> &RoundResults

Results of executing the last successful runtime round.

fn runtime_state(&mut self) -> &mut Self::S

Runtime state store.

fn consensus_state(&self) -> &ConsensusState

Consensus state.

fn epoch(&self) -> EpochTime

Current epoch.

fn tx_auth_info(&self) -> Option<&AuthInfo>

Transaction authentication information.

Only present if this is a transaction processing context.

fn emit_event<E: Event>(&mut self, event: E)

Emits an event.

fn emit_message(
&mut self,
msg: Message,
hook: MessageEventHookInvocation
) -> Result<(), Error>

Attempts to emit consensus runtime message.

fn io_ctx(&self) -> IoContext

Returns a child io_ctx.

fn commit(self) -> (Tags, Vec<(Message, MessageEventHookInvocation)>)

Commit any changes made to storage, return any emitted tags and runtime messages. It consumes the transaction context.

fn value<V: Any>(&mut self, key: &'static str) -> ContextValue<'_, V>

Fetches a value entry associated with the context.

Loading content...
Provided methods

fn is_check_only(&self) -> bool

Whether the transaction is just being checked for validity.

fn is_simulation(&self) -> bool

Whether the transaction is just being simulated.

fn tx_caller_address(&self) -> Option<Address>

Authenticated address of the caller.

In case there are multiple signers of a transaction, this will return the address corresponding to the first signer.

Only present if this is a transaction processing context.

fn emit_messages(
&mut self,
msgs: Vec<(Message, MessageEventHookInvocation)>
) -> Result<(), Error>

Attempts to emit multiple consensus runtime messages.

Loading content...