Trait oasis_runtime_sdk::context::Context

pub trait Context {
    type Runtime: Runtime;

    // Required methods
    fn get_logger(&self, module: &'static str) -> Logger;
    fn mode(&self) -> Mode;
    fn host_info(&self) -> &HostInfo;
    fn key_manager(&self) -> Option<&dyn KeyManager>;
    fn runtime_header(&self) -> &Header;
    fn runtime_round_results(&self) -> &RoundResults;
    fn consensus_state(&self) -> &ConsensusState;
    fn history(&self) -> &dyn HistoryHost;
    fn epoch(&self) -> EpochTime;
    fn emit_event<E: Event>(&mut self, event: E);
    fn emit_etag(&mut self, etag: EventTag);
    fn emit_etags(&mut self, etags: EventTags);
    fn commit(self) -> State;
    fn rollback(self) -> EventTags;
    fn value<V: Any>(&mut self, key: &'static str) -> ContextValue<'_, V>;
    fn remaining_messages(&self) -> u32;
    fn limit_max_messages(&mut self, max_messages: u32) -> Result<(), Error>;
    fn with_child<F, Rs>(&mut self, mode: Mode, f: F) -> Rs
       where F: FnOnce(RuntimeBatchContext<'_, Self::Runtime>) -> Rs;
    fn rng(&mut self, pers: &[u8]) -> Result<LeafRng, Error>;

    // Provided methods
    fn is_check_only(&self) -> bool { ... }
    fn is_pre_schedule(&self) -> bool { ... }
    fn is_simulation(&self) -> bool { ... }
    fn should_execute_contracts(&self) -> bool { ... }
    fn is_allowed_query<R: Runtime>(&self, method: &str) -> bool { ... }
    fn local_config<T>(&self, key: &str) -> Option<T>
       where T: Decode { ... }
    fn runtime_id(&self) -> &Namespace { ... }
    fn is_confidential(&self) -> bool { ... }
    fn with_simulation<F, Rs>(&mut self, f: F) -> Rs
       where F: FnOnce(RuntimeBatchContext<'_, Self::Runtime>) -> Rs { ... }

}

Runtime SDK context.

Required Associated Types

type Runtime: Runtime

Runtime that the context is being invoked in.

Required Methods

fn get_logger(&self, module: &'static str) -> Logger

Returns a logger.

fn mode(&self) -> Mode

Context mode.

fn host_info(&self) -> &HostInfo

Information about the host environment.

fn key_manager(&self) -> Option<&dyn KeyManager>

The key manager, if the runtime is confidential.

fn runtime_header(&self) -> &Header

Last runtime block header.

fn runtime_round_results(&self) -> &RoundResults

Results of executing the last successful runtime round.

fn consensus_state(&self) -> &ConsensusState

Consensus state.

fn history(&self) -> &dyn HistoryHost

Historical state.

fn epoch(&self) -> EpochTime

Current epoch.

fn emit_event<E: Event>(&mut self, event: E)

Emits an event by transforming it into a tag and emitting a tag.

fn emit_etag(&mut self, etag: EventTag)

Emits a tag.

fn emit_etags(&mut self, etags: EventTags)

Emits event tags.

fn commit(self) -> State

Return any emitted tags and runtime messages. It consumes the transaction context.

Storage

This does not commit any storage transaction.

fn rollback(self) -> EventTags

Rollback any changes made by this context. This method only needs to be called explicitly in case you want to retrieve possibly emitted unconditional events. Simply dropping the context without calling commit will also result in a rollback.

Storage

This does not rollback any storage transaction.

fn value<V: Any>(&mut self, key: &'static str) -> ContextValue<'_, V>

Fetches a value entry associated with the context.

fn remaining_messages(&self) -> u32

Number of consensus messages that can still be emitted.

fn limit_max_messages(&mut self, max_messages: u32) -> Result<(), Error>

Set an upper limit on the number of consensus messages that can be emitted in this context. Note that the limit can only be decreased and calling this function will return an error in case the passed max_messages is higher than the current limit.

fn with_child<F, Rs>(&mut self, mode: Mode, f: F) -> Rswhere F: FnOnce(RuntimeBatchContext<'_, Self::Runtime>) -> Rs,

Executes a function in a child context with the given mode.

The context collects its own messages and starts with an empty set of context values.

Storage

This does not start a new storage transaction. Start a transaction and explicitly commit or rollback if you want to discard storage side effects.

fn rng(&mut self, pers: &[u8]) -> Result<LeafRng, Error>

Returns a random number generator, if it is available, with optional personalization.

Provided Methods

fn is_check_only(&self) -> bool

Whether the transaction is just being checked for validity.

fn is_pre_schedule(&self) -> bool

Whether the transaction is just being checked for validity before being scheduled.

fn is_simulation(&self) -> bool

Whether the transaction is just being simulated.

fn should_execute_contracts(&self) -> bool

Whether smart contracts should be executed in this context.

fn is_allowed_query<R: Runtime>(&self, method: &str) -> bool

Whether method is an allowed query per policy in the local config.

fn local_config<T>(&self, key: &str) -> Option<T>where T: Decode,

Returns node operator-provided local configuration.

This method will always return None in Mode::ExecuteTx contexts.

fn runtime_id(&self) -> &Namespace

Runtime ID.

fn is_confidential(&self) -> bool

Whether the context has a key manager available (e.g. the runtime is confidential).

fn with_simulation<F, Rs>(&mut self, f: F) -> Rswhere F: FnOnce(RuntimeBatchContext<'_, Self::Runtime>) -> Rs,

Executes a function in a simulation context.

The simulation context collects its own messages and starts with an empty set of context values.

Storage

This does not start a new storage transaction. Start a transaction and explicitly commit or rollback if you want to discard storage side effects.