Trait oasis_runtime_sdk::context::TxContext

pub trait TxContext: Context {
    fn tx_index(&self) -> usize;

    fn tx_size(&self) -> u32;

    fn tx_auth_info(&self) -> &AuthInfo;

    fn tx_call_format(&self) -> CallFormat;

    fn tx_value<V: Any>(&mut self, key: &'static str) -> ContextValue<'_, V>;

    fn emit_message(
        &mut self, 
        msg: Message, 
        hook: MessageEventHookInvocation
    ) -> Result<(), Error>;

    fn emit_unconditional_event<E: Event>(&mut self, event: E);

    fn tx_caller_address(&self) -> Address { ... }
}

Runtime SDK transaction context.

Required methods

fn tx_index(&self) -> usize

The index of the transaction in the batch.

fn tx_size(&self) -> u32

Transaction size in bytes.

fn tx_auth_info(&self) -> &AuthInfo

Transaction authentication information.

fn tx_call_format(&self) -> CallFormat

The transaction’s call format.

fn tx_value<V: Any>(&mut self, key: &'static str) -> ContextValue<'_, V>

Fetches an entry pointing to a value associated with the transaction.

fn emit_message(
&mut self,
msg: Message,
hook: MessageEventHookInvocation
) -> Result<(), Error>

Emit a consensus message.

fn emit_unconditional_event<E: Event>(&mut self, event: E)

Similar as emit_event but the event will persist even in case the transaction that owns this context fails.

Provided methods

fn tx_caller_address(&self) -> Address

Authenticated address of the caller.

In case there are multiple signers of a transaction, this will return the address corresponding to the first signer.