Trait oasis_runtime_sdk::context::TxContext

pub trait TxContext: Context {
    fn tx_size(&self) -> u32;

    fn tx_auth_info(&self) -> &AuthInfo;

    fn tx_value<V: Any>(&mut self, key: &'static str) -> ContextValue<'_, V>;

    fn emit_message(
        &mut self, 
        msg: Message, 
        hook: MessageEventHookInvocation
    ) -> Result<(), Error>;

    fn tx_caller_address(&self) -> Address { ... }
}

Runtime SDK transaction context.

Required methods

fn tx_size(&self) -> u32

Transaction size in bytes.

fn tx_auth_info(&self) -> &AuthInfo

Transaction authentication information.

fn tx_value<V: Any>(&mut self, key: &'static str) -> ContextValue<'_, V>

Fetches an entry pointing to a value associated with the transaction.

fn emit_message(
&mut self,
msg: Message,
hook: MessageEventHookInvocation
) -> Result<(), Error>

Emit a consensus message.

Provided methods

fn tx_caller_address(&self) -> Address

Authenticated address of the caller.

In case there are multiple signers of a transaction, this will return the address corresponding to the first signer.