oasis_runtime_sdk::dispatcher

Struct Dispatcher

pub struct Dispatcher<R: Runtime> { /* private fields */ }

The runtime dispatcher.

Implementations

impl<R: Runtime> Dispatcher<R>

pub fn decode_tx<C: Context>(ctx: &C, tx: &[u8]) -> Result<Transaction, Error>

Decode a runtime transaction.

pub fn dispatch_tx_call<C: Context>( ctx: &C, call: Call, opts: &DispatchOptions<'_>, ) -> (CallResult, Metadata)

Run the dispatch steps inside a transaction context. This includes the before call hooks, the call itself and after call hooks. The after call hooks are called regardless if the call succeeds or not.

pub fn dispatch_tx_opts<C: Context>( ctx: &C, tx: Transaction, opts: &DispatchOptions<'_>, ) -> Result<DispatchResult, Error>

Dispatch a runtime transaction in the given context with the provided options.

pub fn dispatch_tx<C: Context>( ctx: &C, tx_size: u32, tx: Transaction, tx_index: usize, ) -> Result<DispatchResult, Error>

Dispatch a runtime transaction in the given context.

pub fn check_tx<C: Context>( ctx: &C, tx_size: u32, tx: Transaction, ) -> Result<CheckTxResult, Error>

Check whether the given transaction is valid.

pub fn execute_tx_opts<C: Context>( ctx: &C, tx: Transaction, opts: &DispatchOptions<'_>, ) -> Result<(CallResult, Tags), Error>

Execute the given transaction, returning unserialized results.

pub fn execute_tx<C: Context>( ctx: &C, tx_size: u32, tx_hash: Hash, tx: Transaction, tx_index: usize, ) -> Result<ExecuteTxResult, Error>

Execute the given transaction.

pub fn prefetch_tx( prefixes: &mut BTreeSet<Prefix>, tx: Transaction, ) -> Result<(), RuntimeError>

Prefetch prefixes for the given transaction.

pub fn dispatch_query<C: Context>( ctx: &C, method: &str, args: Vec<u8>, ) -> Result<Vec<u8>, RuntimeError>

Process the given runtime query.

pub fn register_enclaverpc(&self, rpc: &mut RpcDispatcher)
where R: Runtime + Send + Sync + 'static,

Register EnclaveRPC methods.