Trait oasis_runtime_sdk::error::Error

pub trait Error: Error {
    fn module_name(&self) -> &str;

    fn code(&self) -> u32;

    fn to_call_result(&self) -> CallResult { ... }
}

A runtime error that gets propagated to the caller.

It extends std::error::Error with module name and error code so that errors can be easily serialized and transferred between different processes.

This trait can be derived:

const MODULE_NAME: &str = "my-module";
#[derive(Clone, Debug, Error, thiserror::Error)]
#[sdk_error(autonumber)] // `module_name` meta is required if `MODULE_NAME` isn't in scope
enum Error {
   #[error("invalid argument")]
   InvalidArgument,          // autonumbered to 0

   #[error("forbidden")]
   #[sdk_error(code = 401)]  // manually numbered to 403 (`code` or autonumbering is required)
   Forbidden,
}

Required methods

fn module_name(&self) -> &str

Name of the module that emitted the error.

fn code(&self) -> u32

Error code uniquely identifying the error.

Loading content...
Provided methods

fn to_call_result(&self) -> CallResult

Converts the error into a call result.

Loading content...