oasis_runtime_sdk::history

Trait HistoryHost

pub trait HistoryHost {
    // Required methods
    fn consensus_state_at(&self, height: u64) -> Result<ConsensusState, Error>;
    fn consensus_events_at(
        &self,
        height: u64,
        kind: EventKind,
    ) -> Result<Vec<Event>, Error>;
}

Interface to the runtime host to fetch historic information.

Required Methods

fn consensus_state_at(&self, height: u64) -> Result<ConsensusState, Error>

Fetch historic consensus state after executing the block at given height.

fn consensus_events_at( &self, height: u64, kind: EventKind, ) -> Result<Vec<Event>, Error>

Fetch events emitted during execution of the block at given height.