Trait oasis_runtime_sdk::module::BlockHandler

pub trait BlockHandler {
    fn begin_block(_ctx: &mut DispatchContext<'_>) { ... }

    fn end_block(_ctx: &mut DispatchContext<'_>) { ... }
}

Block handler.

Provided methods

fn begin_block(_ctx: &mut DispatchContext<'_>)

Perform any common actions at the start of the block (before any transactions have been executed).

fn end_block(_ctx: &mut DispatchContext<'_>)

Perform any common actions at the end of the block (after all transactions have been executed).

Loading content...