Trait oasis_runtime_sdk::module::BlockHandler

pub trait BlockHandler {
    // Provided methods
    fn begin_block<C: Context>(_ctx: &C) { ... }
    fn end_block<C: Context>(_ctx: &C) { ... }
}

Block handler.

Provided Methods

fn begin_block<C: Context>(_ctx: &C)

Perform any common actions at the start of the block (before any transactions have been executed).

fn end_block<C: Context>(_ctx: &C)

Perform any common actions at the end of the block (after all transactions have been executed).

Object Safety

This trait is not object safe.