Trait oasis_runtime_sdk::module::MethodHandler

pub trait MethodHandler {
    fn prefetch(
        _prefixes: &mut BTreeSet<Prefix>,
        _method: &str,
        body: Value,
        _auth_info: &AuthInfo
    ) -> DispatchResult<Value, Result<(), RuntimeError>> { ... }
    fn dispatch_call<C: TxContext>(
        _ctx: &mut C,
        _method: &str,
        body: Value
    ) -> DispatchResult<Value, CallResult> { ... }
    fn dispatch_query<C: Context>(
        _ctx: &mut C,
        _method: &str,
        args: Value
    ) -> DispatchResult<Value, Result<Value, RuntimeError>> { ... }
    fn dispatch_message_result<C: Context>(
        _ctx: &mut C,
        _handler_name: &str,
        result: MessageResult
    ) -> DispatchResult<MessageResult, ()> { ... }
    fn supported_methods() -> Vec<MethodHandlerInfo> { ... }
    fn is_expensive_query(_method: &str) -> bool { ... }
    fn is_allowed_private_km_query(_method: &str) -> bool { ... }
    fn is_allowed_interactive_call(_method: &str) -> bool { ... }
}

Method handler.

Provided Methods

fn prefetch(
_prefixes: &mut BTreeSet<Prefix>,
_method: &str,
body: Value,
_auth_info: &AuthInfo
) -> DispatchResult<Value, Result<(), RuntimeError>>

Add storage prefixes to prefetch.

fn dispatch_call<C: TxContext>(
_ctx: &mut C,
_method: &str,
body: Value
) -> DispatchResult<Value, CallResult>

Dispatch a call.

fn dispatch_query<C: Context>(
_ctx: &mut C,
_method: &str,
args: Value
) -> DispatchResult<Value, Result<Value, RuntimeError>>

Dispatch a query.

fn dispatch_message_result<C: Context>(
_ctx: &mut C,
_handler_name: &str,
result: MessageResult
) -> DispatchResult<MessageResult, ()>

Dispatch a message result.

fn supported_methods() -> Vec<MethodHandlerInfo>

Lists the names of all RPC methods exposed by this module. The result is informational only. An empty return vector means that the implementor does not care to list the methods, or the implementor is a tuple of modules.

fn is_expensive_query(_method: &str) -> bool

Checks whether the given query method is tagged as expensive.

fn is_allowed_private_km_query(_method: &str) -> bool

Checks whether the given query is allowed to access private key manager state.

fn is_allowed_interactive_call(_method: &str) -> bool

Checks whether the given call is allowed to be called interactively via read-only transactions.