Trait oasis_runtime_sdk::modules::consensus::API

pub trait API {
    fn transfer<C: TxContext>(
        ctx: &mut C,
        to: Address,
        amount: &BaseUnits,
        hook: MessageEventHookInvocation
    ) -> Result<(), Error>;
    fn withdraw<C: TxContext>(
        ctx: &mut C,
        from: Address,
        amount: &BaseUnits,
        hook: MessageEventHookInvocation
    ) -> Result<(), Error>;
    fn escrow<C: TxContext>(
        ctx: &mut C,
        to: Address,
        amount: &BaseUnits,
        hook: MessageEventHookInvocation
    ) -> Result<(), Error>;
    fn reclaim_escrow<C: TxContext>(
        ctx: &mut C,
        from: Address,
        amount: u128,
        hook: MessageEventHookInvocation
    ) -> Result<(), Error>;
    fn consensus_denomination<C: Context>(
        ctx: &mut C
    ) -> Result<Denomination, Error>;
    fn ensure_compatible_tx_signer<C: TxContext>(ctx: &C) -> Result<(), Error>;
    fn account<C: Context>(
        ctx: &C,
        addr: Address
    ) -> Result<ConsensusAccount, Error>;
    fn amount_from_consensus<C: Context>(
        ctx: &mut C,
        amount: u128
    ) -> Result<u128, Error>;
    fn amount_to_consensus<C: Context>(
        ctx: &mut C,
        amount: u128
    ) -> Result<u128, Error>;
}

Interface that can be called from other modules.

Required Methods

fn transfer<C: TxContext>(
ctx: &mut C,
to: Address,
amount: &BaseUnits,
hook: MessageEventHookInvocation
) -> Result<(), Error>

Transfer an amount from the runtime account.

fn withdraw<C: TxContext>(
ctx: &mut C,
from: Address,
amount: &BaseUnits,
hook: MessageEventHookInvocation
) -> Result<(), Error>

Withdraw an amount into the runtime account.

fn escrow<C: TxContext>(
ctx: &mut C,
to: Address,
amount: &BaseUnits,
hook: MessageEventHookInvocation
) -> Result<(), Error>

Escrow an amount of the runtime account funds.

fn reclaim_escrow<C: TxContext>(
ctx: &mut C,
from: Address,
amount: u128,
hook: MessageEventHookInvocation
) -> Result<(), Error>

Reclaim an amount of runtime staked shares.

fn consensus_denomination<C: Context>(
ctx: &mut C
) -> Result<Denomination, Error>

Returns consensus token denomination.

fn ensure_compatible_tx_signer<C: TxContext>(ctx: &C) -> Result<(), Error>

Ensures transaction signer is consensus compatible.

fn account<C: Context>(ctx: &C, addr: Address) -> Result<ConsensusAccount, Error>

Query consensus account info.

fn amount_from_consensus<C: Context>(
ctx: &mut C,
amount: u128
) -> Result<u128, Error>

Convert runtime amount to consensus amount, scaling as needed.

fn amount_to_consensus<C: Context>(
ctx: &mut C,
amount: u128
) -> Result<u128, Error>

Convert consensus amount to runtime amount, scaling as needed.