Trait oasis_runtime_sdk::modules::consensus_accounts::API

pub trait API {
    fn deposit<C: TxContext>(
        ctx: &mut C,
        from: Address,
        nonce: u64,
        to: Address,
        amount: BaseUnits
    ) -> Result<(), Error>;
    fn withdraw<C: TxContext>(
        ctx: &mut C,
        from: Address,
        nonce: u64,
        to: Address,
        amount: BaseUnits
    ) -> Result<(), Error>;
    fn delegate<C: TxContext>(
        ctx: &mut C,
        from: Address,
        nonce: u64,
        to: Address,
        amount: BaseUnits
    ) -> Result<(), Error>;
    fn undelegate<C: TxContext>(
        ctx: &mut C,
        from: Address,
        nonce: u64,
        to: Address,
        shares: u128
    ) -> Result<(), Error>;
}

Interface that can be called from other modules.

Required Methods

fn deposit<C: TxContext>(
ctx: &mut C,
from: Address,
nonce: u64,
to: Address,
amount: BaseUnits
) -> Result<(), Error>

Transfer from consensus staking account to runtime account.

Arguments

nonce: A caller-provided sequence number that will help identify the success/fail events. When called from a deposit transaction, we use the signer nonce.

fn withdraw<C: TxContext>(
ctx: &mut C,
from: Address,
nonce: u64,
to: Address,
amount: BaseUnits
) -> Result<(), Error>

Transfer from runtime account to consensus staking account.

Arguments

nonce: A caller-provided sequence number that will help identify the success/fail events. When called from a withdraw transaction, we use the signer nonce.

fn delegate<C: TxContext>(
ctx: &mut C,
from: Address,
nonce: u64,
to: Address,
amount: BaseUnits
) -> Result<(), Error>

Delegate from runtime account to consensus staking account.

Arguments

nonce: A caller-provided sequence number that will help identify the success/fail events. When called from a delegate transaction, we use the signer nonce.

fn undelegate<C: TxContext>(
ctx: &mut C,
from: Address,
nonce: u64,
to: Address,
shares: u128
) -> Result<(), Error>

Start the undelegation process of the given number of shares from consensus staking account to runtime account.

Arguments

nonce: A caller-provided sequence number that will help identify the success/fail events. When called from an undelegate transaction, we use the signer nonce.