Module oasis_runtime_sdk::modules::consensus_accounts::types

Consensus module types.

Structs

AccountBalance

BalanceQuery

Balance query.

ConsensusAccountQuery

Consensus account query.

ConsensusDelegateContext

Context for consensus delegate message handler.

ConsensusError

Error details from the consensus layer.

ConsensusTransferContext

Context for consensus transfer message handler.

ConsensusUndelegateContext

Context for consensus undelegate message handler.

ConsensusWithdrawContext

Context for consensus withdraw message handler.

Delegate

Delegate from runtime call.

DelegationInfo

Information about a delegation.

DelegationQuery

Delegation query.

DelegationsQuery

Delegations query.

Deposit

Deposit into runtime call. Transfer from consensus staking to an account in this runtime. The transaction signer has a consensus layer allowance benefiting this runtime’s staking address. The to address runtime account gets the tokens.

ExtendedDelegationInfo

Extended information about a delegation.

Receipt

A receipt.

TakeReceipt

Take receipt internal runtime call.

Undelegate

Undelegate into runtime call.

UndelegationInfo

Information about an undelegation.

UndelegationsQuery

Undelegations query.

Withdraw

Withdraw from runtime call. Transfer from an account in this runtime to consensus staking. The to address consensus staking account gets the tokens.

Enums

ReceiptKind

Kind of receipt.