Trait oasis_runtime_sdk::modules::core::API

pub trait API {
    fn use_gas(ctx: &mut TxContext<'_, '_>, gas: u64) -> Result<(), Error>;
}

Required methods

fn use_gas(ctx: &mut TxContext<'_, '_>, gas: u64) -> Result<(), Error>

Attempt to use gas. Gas limits are per-batch (max_batch_gas in Parameters) and per-transaction (.ai.fee.gas). If the gas specified would cause either total used to exceed its limit, fails with Error::OutOfGas or Error::BatchOutOfGas, and neither gas usage is increased. Per-transaction gas is not assessed when C is not a transaction processing context.

Loading content...