oasis_runtime_sdk::modules::core

Trait Config

pub trait Config: 'static {
    const DEFAULT_LOCAL_MIN_GAS_PRICE: Lazy<BTreeMap<Denomination, u128>> = _;
    const DEFAULT_LOCAL_ESTIMATE_GAS_SEARCH_MAX_ITERS: u64 = 0u64;
    const ESTIMATE_GAS_EXTRA_FAIL: Lazy<BTreeMap<&'static str, u64>> = _;
    const MIN_GAS_PRICE_EXEMPT_METHODS: Lazy<BTreeSet<&'static str>> = _;
    const EMIT_GAS_USED_EVENTS: bool = true;
    const ALLOW_INTERACTIVE_READ_ONLY_TRANSACTIONS: bool = false;
    const GAS_COST_CALL_CALLDATA_PUBLIC_KEY: u64 = 20u64;
    const GAS_COST_CALL_CURRENT_EPOCH: u64 = 10u64;
}

Module configuration.

Provided Associated Constants

const DEFAULT_LOCAL_MIN_GAS_PRICE: Lazy<BTreeMap<Denomination, u128>> = _

Default local minimum gas price configuration that is used in case no overrides are set in local per-node configuration.

const DEFAULT_LOCAL_ESTIMATE_GAS_SEARCH_MAX_ITERS: u64 = 0u64

Default local estimate gas max search iterations configuration that is used in case no overrides are set in the local per-node configuration.

const ESTIMATE_GAS_EXTRA_FAIL: Lazy<BTreeMap<&'static str, u64>> = _

Estimated gas amount to be added to failed transaction simulations for selected methods.

const MIN_GAS_PRICE_EXEMPT_METHODS: Lazy<BTreeSet<&'static str>> = _

Methods which are exempt from minimum gas price requirements.

const EMIT_GAS_USED_EVENTS: bool = true

Whether gas used events should be emitted for every transaction.

Confidential runtimes may want to disable this as it is a possible side channel.

const ALLOW_INTERACTIVE_READ_ONLY_TRANSACTIONS: bool = false

Whether to allow submission of read-only transactions in an interactive way.

Note that execution of such transactions is allowed to access confidential state.

const GAS_COST_CALL_CALLDATA_PUBLIC_KEY: u64 = 20u64

The gas cost of the internal call to retrieve the current calldata public key.

const GAS_COST_CALL_CURRENT_EPOCH: u64 = 10u64

The gas cost of the internal call to retrieve the current epoch.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.