oasis_runtime_sdk::modules::rofl

Trait Config

pub trait Config: 'static {

    const GAS_COST_CALL_CREATE: u64 = 100_000u64;
    const GAS_COST_CALL_UPDATE: u64 = 100_000u64;
    const GAS_COST_CALL_REMOVE: u64 = 10_000u64;
    const GAS_COST_CALL_REGISTER: u64 = 100_000u64;
    const GAS_COST_CALL_IS_AUTHORIZED_ORIGIN: u64 = 1_000u64;
    const GAS_COST_CALL_AUTHORIZED_ORIGIN_NODE: u64 = 2_000u64;
    const GAS_COST_CALL_AUTHORIZED_ORIGIN_ENTITY: u64 = 2_000u64;
    const GAS_COST_CALL_STAKE_THRESHOLDS: u64 = 10u64;
    const GAS_COST_CALL_DERIVE_KEY: u64 = 10_000u64;
    const MAX_METADATA_PAIRS: usize = 64usize;
    const MAX_METADATA_KEY_SIZE: usize = 1_024usize;
    const MAX_METADATA_VALUE_SIZE: usize = 16_384usize;
    const STAKE_APP_CREATE: BaseUnits = _;
    const DERIVE_KEY_MAX_KEY_ID_LENGTH: usize = 128usize;

}

Module configuration.

Provided Associated Constants

const GAS_COST_CALL_CREATE: u64 = 100_000u64

Gas cost of rofl.Create call.

const GAS_COST_CALL_UPDATE: u64 = 100_000u64

Gas cost of rofl.Update call.

const GAS_COST_CALL_REMOVE: u64 = 10_000u64

Gas cost of rofl.Remove call.

const GAS_COST_CALL_REGISTER: u64 = 100_000u64

Gas cost of rofl.Register call.

const GAS_COST_CALL_IS_AUTHORIZED_ORIGIN: u64 = 1_000u64

Gas cost of rofl.IsAuthorizedOrigin call.

const GAS_COST_CALL_AUTHORIZED_ORIGIN_NODE: u64 = 2_000u64

Gas cost of rofl.AuthorizedOriginNode call.

const GAS_COST_CALL_AUTHORIZED_ORIGIN_ENTITY: u64 = 2_000u64

Gas cost of rofl.AuthorizedOriginEntity call.

const GAS_COST_CALL_STAKE_THRESHOLDS: u64 = 10u64

Gas cost of rofl.StakeThresholds call.

const GAS_COST_CALL_DERIVE_KEY: u64 = 10_000u64

Gas cost of rofl.DeriveKey call.

const MAX_METADATA_PAIRS: usize = 64usize

Maximum number of metadata key-value pairs.

const MAX_METADATA_KEY_SIZE: usize = 1_024usize

Maximum metadata key size.

const MAX_METADATA_VALUE_SIZE: usize = 16_384usize

Maximum metadata value size.

const STAKE_APP_CREATE: BaseUnits = _

Amount of stake required for maintaining an application.

The stake is held in escrow and is returned to the administrator when the application is removed.

const DERIVE_KEY_MAX_KEY_ID_LENGTH: usize = 128usize

Maximum key identifier length for rofl.DeriveKey call.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.