Trait oasis_runtime_sdk::runtime::Runtime

pub trait Runtime {
    type Core: API;
    type Modules: AuthHandler + MigrationHandler + MethodHandler + BlockHandler + InvariantHandler;

    const VERSION: Version;
    const STATE_VERSION: u32;
    const PREFETCH_LIMIT: u16;
    const SCHEDULE_CONTROL: Option<ScheduleControl>;

    fn genesis_state() -> <Self::Modules as MigrationHandler>::Genesis;

    fn trusted_policy_signers() -> Option<TrustedPolicySigners> { ... }

    fn consensus_trust_root() -> Option<TrustRoot> { ... }

    fn migrate_state<C: Context>(_ctx: &mut C) { ... }

    fn migrate<C: Context>(ctx: &mut C) { ... }

    fn start()
    where
        Self: Sized + Send + Sync + 'static,
    { ... }
}

A runtime.

Associated Types

type Core: API

Module that provides the core API.

type Modules: AuthHandler + MigrationHandler + MethodHandler + BlockHandler + InvariantHandler

Supported modules.

Associated Constants

const VERSION: Version

Runtime version.

const STATE_VERSION: u32

State version.

const PREFETCH_LIMIT: u16

Prefetch limit. To enable prefetch set it to a non-zero value.

const SCHEDULE_CONTROL: Option<ScheduleControl>

Whether the runtime should take control of transaction scheduling.

Required methods

fn genesis_state() -> <Self::Modules as MigrationHandler>::Genesis

Genesis state for the runtime.

Provided methods

fn trusted_policy_signers() -> Option<TrustedPolicySigners>

Return the trusted policy signers for this runtime; if None, a key manager connection will not be established on startup.

fn consensus_trust_root() -> Option<TrustRoot>

Return the consensus layer trust root for this runtime; if None, consensus layer integrity verification will not be performed.

fn migrate_state<C: Context>(_ctx: &mut C)

Perform runtime-specific state migration. This method is only called when the recorded state version does not match STATE_VERSION.

fn migrate<C: Context>(ctx: &mut C)

Perform state migrations if required.

fn start() where
Self: Sized + Send + Sync + 'static,

Start the runtime.