oasis_runtime_sdk::schedule_control

Trait ScheduleControlHost

pub trait ScheduleControlHost: Send + Sync {
    // Required method
    fn fetch_tx_batch(
        &self,
        offset: Option<Hash>,
        limit: u32,
    ) -> Result<Option<TxnBatch>, Error>;
}

Interface to the runtime host that supports schedule control features.

Required Methods

fn fetch_tx_batch( &self, offset: Option<Hash>, limit: u32, ) -> Result<Option<TxnBatch>, Error>

Fetch the specified set of transactions from the host’s transaction queue.

Offset specifies the transaction hash that should serve as an offset when returning transactions from the pool. Transactions will be skipped until the given hash is encountered and only following transactions will be returned.