Module oasis_runtime_sdk::storage

Storage.

Re-exports

pub use confidential::ConfidentialStore;

pub use confidential::Error as ConfidentialStoreError;

Modules

confidential

Structs

HashedStore

A key-value store that hashes all keys and stores them as H(k) || k.

HostStore

A store for a specific state root that talks to the runtime host.

MKVSStore

A key-value store backed by MKVS.

OverlayStore

An overlay store which keeps values locally until explicitly committed.

Prefix

A key prefix.

PrefixStore

A key-value store that prefixes all keys with the given prefix.

TypedStore

A key-value store that transparently handles serialization/deserialization.

Traits

NestedStore

A key-value store that supports the commit operation.

Store

A key-value store.