oasis_runtime_sdk::storage::host

Function new_mkvs_tree_for_round

pub async fn new_mkvs_tree_for_round(
    host: Arc<Protocol>,
    consensus_verifier: &Arc<dyn Verifier>,
    runtime_id: Namespace,
    round: u64,
    root_type: RootType,
) -> Result<Tree>

Create a new MKVS tree for the given host and runtime at the given round.

The corresponding root hash is fetched by looking it up in consensus layer state, verified by the passed verifier to be correct.