oasis_runtime_sdk::storage

Trait NestedStore

pub trait NestedStore: Store {
    type Inner;

    // Required methods
    fn commit(self) -> Self::Inner;
    fn rollback(self) -> Self::Inner;
    fn has_pending_updates(&self) -> bool;
    fn pending_update_byte_size(&self) -> usize;
}

A key-value store that supports the commit operation.

Required Associated Types

type Inner

Type of the inner store.

Required Methods

fn commit(self) -> Self::Inner

Commit any changes to the underlying store.

If this method is not called the changes may be discarded by the store.

fn rollback(self) -> Self::Inner

Rollback any changes.

fn has_pending_updates(&self) -> bool

Whether there are any store updates pending to be committed.

fn pending_update_byte_size(&self) -> usize

Size (in bytes) of any pending updates.